commit: expanded information on orders, added conclude option for orders to lock it from changes (order _items not affected by said "lock" yet)

+++ main.py
@@ -36,6 +36,7 @@ def init_db():
                 id INTEGER PRIMARY KEY AUTOINCREMENT,
                 customer_id INTEGER,
                 order_date TEXT,
+                purchase_done INTEGER DEFAULT 0,
                 FOREIGN KEY (customer_id) REFERENCES customers (id)
             );
 
@@ -161,9 +162,16 @@ def list_orders():
     with sqlite3.connect(DATABASE) as conn:
         cur = conn.cursor()
         cur.execute('''
-            SELECT orders.id, customers.name, orders.order_date
+            SELECT orders.id, customers.name, orders.order_date, total_value, orders.purchase_done
             FROM orders
             LEFT JOIN customers ON orders.customer_id = customers.id
+            LEFT JOIN (
+                    SELECT orit.order_id, SUM(p.price * orit.quantity) AS total_value
+                    FROM order_items AS orit
+                    LEFT JOIN products AS p
+                    ON orit.product_id = p.id
+                    GROUP BY orit.order_id
+                ) AS total_value ON orders.id = total_value.order_id
         ''')
         orders = cur.fetchall()
     return render_template('orders/list.html', orders=orders)
@@ -193,15 +201,31 @@ def edit_order(order_id):
             return redirect(url_for('list_orders'))
         cur.execute('SELECT * FROM orders WHERE id=?', (order_id,))
         order = cur.fetchone()
+        
+        # Checks if the order was already concluded, to not allow it to be edited if it was
+        cur.execute('SELECT purchase_done FROM orders WHERE id=?', (order_id,))
+        purchase_done = cur.fetchone()
+        if purchase_done[0] == 1:
+            return redirect(url_for('list_orders'))
+        
+        # Fetch customers for the dropdown
         cur.execute('SELECT * FROM customers')
         customers = cur.fetchall()
     return render_template('orders/edit.html', order=order, customers=customers)
+
+@app.route('/orders/<int:order_id>/conclude')
+def conclude_order(order_id):
+    with sqlite3.connect(DATABASE) as conn:
+        cur = conn.cursor()
+        cur.execute('UPDATE orders SET purchase_done = 1 WHERE id=?', (order_id,))
+        conn.commit()
+    return redirect(url_for('list_orders'))
 
 @app.route('/orders/<int:order_id>/delete')
 def delete_order(order_id):
     with sqlite3.connect(DATABASE) as conn:
         cur = conn.cursor()
-        cur.execute('DELETE FROM orders WHERE id=?', (order_id,))
+        cur.execute('DELETE FROM orders WHERE id=? AND purchase_done = 0', (order_id,))
         conn.commit()
     return redirect(url_for('list_orders'))
 
@@ -214,13 +238,28 @@ def list_order_items(order_id):
     with sqlite3.connect(DATABASE) as conn:
         cur = conn.cursor()
         cur.execute('''
-            SELECT order_items.id, order_items.order_id, products.name, order_items.quantity
+            SELECT order_items.id, order_items.order_id, products.name, order_items.quantity, products.price, (products.price * order_items.quantity) AS item_total_value
             FROM order_items
             LEFT JOIN products ON order_items.product_id = products.id
             WHERE order_items.order_id = {0}
         '''.format(order_id))
         order_items = cur.fetchall()
-    return render_template('order_items/list.html', order_items=order_items, order_id=order_id)
+        
+        cur.execute('''
+            SELECT SUM(p.price * orit.quantity) AS total_value
+            FROM order_items AS orit
+            LEFT JOIN products AS p
+            ON orit.product_id = p.id
+            WHERE orit.order_id = {0}
+            GROUP BY orit.order_id
+        '''.format(order_id))
+        total_value = cur.fetchone()
+        
+        # Checks if the order was already concluded, and pass the information forward
+        # so the template won't allow it to be edited if it was concluded
+        cur.execute('SELECT purchase_done FROM orders WHERE id=?', (order_id,))
+        purchase_done = cur.fetchone()
+    return render_template('order_items/list.html', order_items=order_items, order_id=order_id, total_value = (total_value[0] if total_value else 0), order_status=purchase_done[0])
 
 @app.route('/order_items/<int:order_id>/create', methods=['GET', 'POST'])
 def create_order_item(order_id):
@@ -253,7 +292,10 @@ def edit_order_item(item_id):
     with sqlite3.connect(DATABASE) as conn:
         cur = conn.cursor()
         if request.method == 'POST':
-            order_id = request.form['order_id']
+            cur.execute('''SELECT order_id
+                        FROM order_items 
+                        WHERE id=?''', (item_id,))
+            order_id = cur.fetchone()[0]
             product_id = request.form['product_id']
             quantity = request.form['quantity']
             cur.execute('''
@@ -261,7 +303,7 @@ def edit_order_item(item_id):
                 WHERE id=?
             ''', (order_id, product_id, quantity, item_id))
             conn.commit()
-            return redirect(url_for('list_order_items'))
+            return redirect(url_for('list_order_items', order_id=order_id))
 
         cur.execute('SELECT * FROM order_items WHERE id=?', (item_id,))
         order_item = cur.fetchone()
@@ -275,9 +317,26 @@ def edit_order_item(item_id):
 def delete_order_item(item_id):
     with sqlite3.connect(DATABASE) as conn:
         cur = conn.cursor()
+
+        # Checks if the order was already concluded, to not allow it to be edited if it was
+        cur.execute('''SELECT purchase_done, o.id 
+                    FROM order_items AS orit
+                    INNER JOIN orders AS o
+                    ON orit.order_id = o.id
+                    WHERE orit.id = {0}
+                    '''.format(item_id));
+        purchase_done = cur.fetchone()
+
+        if purchase_done[0] == 1:
+            return redirect(url_for('list_orders'))
+        
+        order_id = purchase_done[1]
+
+
+        # If the order was not concluded yet, delete the order item
         cur.execute('DELETE FROM order_items WHERE id=?', (item_id,))
         conn.commit()
-    return redirect(url_for('list_order_items'))
+    return redirect(url_for('list_order_items', order_id=order_id))
 
 # ----------------------------------------
 # ROUTES FOR CATEGORIES
